requirements.txt
hydra-core>=1.3.2
omegaconf>=2.3.0
simpy>=4.1.1
plotly>=5.24.0
pandas>=2.2.0
streamlit>=1.40.0
